Add instructorId parameter to Course.Create method
The `Create` static factory method in the `Course` class has been
modified to include an additional parameter, `instructorId`, of
type `Guid`. The method now creates a `Course` object and assigns
an instructor to it using the new `AssignInstructor` method before
returning the `Course` object. The `AssignInstructor` method is
called within the `Create` method to associate the provided
`instructorId` with the newly created `Course` object.