Remove default values for SortBy and SortOrder properties
The default values for the `SortBy` and `SortOrder` properties in
the `EndpointFilter` record have been removed. Previously, `SortBy`
defaulted to "Id" and `SortOrder` defaulted to "asc". Now, both
properties are nullable strings without default values.

Correspondingly, in the `EndpointFilterModelBinder` class, the
logic for setting `sortBy` and `sortOrder` has been updated to
remove the default values. Previously, if `sortBy` was not
provided in the query, it defaulted to "Id", and if `sortOrder`
was not provided, it defaulted to "asc". Now, these values will
be null if not provided in the query.