Add Duration to Lesson; enhance repositories
Added a `Duration` property of type `TimeSpan` to the `Lesson` class in `Lesson.cs`. Created a new migration `20250102131206_AddTimeSpanToLesson` to add the `Duration` column to the `Lessons` table in the database. Updated the `ApplicationDbContextModelSnapshot.cs` to include the new `Duration` property in the `Lessons` table.

Modified `EnrollmentRepository.cs`, `LessonCategoryRepository.cs`, and `LessonRepository.cs` to include methods for retrieving, adding, and managing entities with filtering, sorting, and paginating capabilities. Updated the `AddAsync` method in `LessonRepository.cs` to update the `TotalDuration` of the course when a new lesson is added.