Change avatar endpoint to POST for uploading avatars
The HTTP method for the `avatar` endpoint in `BlobController.cs` has been changed from `HttpGet` to `HttpPost` to support avatar uploads. The method name has been updated from `GetAvatar` to `UploadAvatar` to reflect this new functionality. The `Authorize` attribute now specifies roles for authorization. The method now uploads an avatar file to the Minio service and returns a `TypedResults.Redirect` to indicate a redirection after the upload operation.