Comment out port 9000 mapping for MinIO in docker-compose
The port mapping for port 9000 has been commented out in the
`docker-compose.yml` file for the MinIO service. Previously,
port 9000 on the host was mapped to port 9000 on the container.
This line has been commented out, and now only port 9001 on the
host is mapped to port 9001 on the container.